layout/generic/test/test_bug597333.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=597333
-->
<head>
  <title>Test for Bug 597333</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

</head>
<body>
<p><a target="_blank" href="https://bugzilla.mozilla.org/show_bug?id=597333">Mozilla Bug 597333</a>

<textarea id="a">&#x202E;</textarea>

<pre id="test">
<script>

function runTests() {
  var t = document.getElementById("a");
  t.focus();
  setTimeout(function() {
    synthesizeKey('VK_RIGHT', {});
    ok(true, "The browser did not crash!");

    SimpleTest.finish();
  }, 0);
}

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(runTests);
</script>
</pre>

</body>
</html>
